Create package application and move PotDAGame and MyXMLReader there
