+++ android/src/com/pottda/game/AndroidLauncher.java
@@ -4,6 +4,7 @@ import android.os.Bundle;
 
 import com.badlogic.gdx.backends.android.AndroidApplication;
 import com.badlogic.gdx.backends.android.AndroidApplicationConfiguration;
+import com.pottda.game.application.PoTDAGame;
 
 public class AndroidLauncher extends AndroidApplication {
 
@@ -13,7 +14,7 @@ public class AndroidLauncher extends AndroidApplication {
         AndroidApplicationConfiguration config = new AndroidApplicationConfiguration();
         config.useAccelerometer = false;
         config.useCompass = false;
-        initialize(new PoTDA(), config);
+        initialize(new PoTDAGame(), config);
     }
 
 }
