+++ desktop/src/com/pottda/game/desktop/DesktopLauncher.java
@@ -2,13 +2,13 @@ package com.pottda.game.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.pottda.game.PoTDA;
+import com.pottda.game.application.PoTDAGame;
 
 class DesktopLauncher {
 	public static void main (String[] arg) {
 		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
         config.width = 800;
         config.height = 480;
-		new LwjglApplication(new PoTDA(), config);
+		new LwjglApplication(new PoTDAGame(), config);
 	}
 }
